pyproject.toml
[tool.poetry]
name = "sopa-de-palavras-api-rest"
version = "0.1.0"
description = "Back and front single repository for sopa-de-palavras"
authors = ["Rodrigo Almeida <c.almeidarodrigo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
